First GUI files and status
all files related to making the gui function are attached.  these have some basic functionality. i'm still working on the business logic for the gui.  otherwise the gui is straightforward and I believe it is easy to use as of right now.
Upon the logic being completed for the current 'set', the other set is nearly identical and will allow for ease of use thanks to Copy-and-Paste!